[#511] SQS: Refactor `SqsReceiver` to use subscription factory convention
