+++ infrastructures/aws-sqs/src/main/java/io/atleon/aws/sqs/BatchRequestFailedException.java
@@ -1,0 +1,12 @@
+package io.atleon.aws.sqs;
+
+import software.amazon.awssdk.services.sqs.model.BatchResultErrorEntry;
+
+import java.util.List;
+
+public class BatchRequestFailedException extends RuntimeException {
+
+    BatchRequestFailedException(String type, List<BatchResultErrorEntry> entries) {
+        super(String.format("Batch request failed! type=%s errors=%s", type, entries));
+    }
+}
